AdDetectorModel/utils/requirements.txt
youtube-dl
ffmpeg-python
scenedetect
webvtt-py
nltk